update version and add error state handling
The `PATCH` version in `version.properties` was updated from 1478 to 1480.
Added an `isError` LiveData to `SelectorViewModel` to handle error states.
Updated `SelectDocumentTypeFragment` and `SelectCatalogTypeFragment` to show `cardNoData` when `isError` is true and to remove unnecessary empty list checks.
Added logic to the `SelectorViewModel` to set the `isError` when the received data is empty.
The view will trigger a reconnect on error.